requirements.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.12.0
transformers>=4.24.0
accelerate>=0.15.0
datasets>=2.8.0
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.5.0
scipy>=1.8.0
tqdm>=4.64.0
opencv-python>=4.6.0
pycocotools>=2.0.4
einops>=0.6.0
wandb>=0.13.0
lpips>=0.1.4
gdown>=4.6.0
requests>=2.28.0
jupyterlab>=3.4.0  # For academic exploration
scikit-image>=0.19.0  # For image processing
seaborn>=0.12.0  # For visualization 